+++ src/main/java/org/folio/services/domainevent/DomainEventType.java
@@ -1,5 +1,5 @@
 package org.folio.services.domainevent;
 
 public enum DomainEventType {
-  UPDATE, DELETE, CREATE
+  UPDATE, DELETE, CREATE, DELETE_ALL
 }