+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Status date set on newly created instance records (MODINVSTOR-509)
 * Item status date is set for newly created item records (MODINVSTOR-508)
+* Allow generation of HRID values with and without leading zeroes (MODINVSTOR-661)
 
 ## 19.4.0 2020-10-08